requirements.txt
asgiref==3.9.1
attrs==25.3.0
cachetools==4.2.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==43.0.3
Django==5.2.6
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
google-api-core==2.10.2
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.30.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
oauth2client==4.1.3
oauthlib==3.3.1
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDrive2==1.21.3
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
